make cards responsive

+++ App/Screens/MyTrips/TripsOverviewScreen.js
@@ -1,7 +1,13 @@
-/* eslint-disable no-alert */
 /* eslint-disable react-native/no-inline-styles */
 import React from 'react';
-import {Text, View, TouchableOpacity, FlatList, Platform} from 'react-native';
+import {
+  Text,
+  View,
+  Alert,
+  TouchableOpacity,
+  FlatList,
+  Platform,
+} from 'react-native';
 import {useSelector, useDispatch} from 'react-redux';
 import {HeaderButtons, Item} from 'react-navigation-header-buttons';
 
@@ -43,7 +49,23 @@ const TripsOverviewScreen = (props) => {
                 alignItems: 'center',
               }}
               onPress={() => {
-                dispatch(tripActions.deleteTrip(itemData.item.id));
+                Alert.alert(
+                  'Delete a trip',
+                  'Are you sure?',
+                  [
+                    {
+                      text: 'Cancel',
+                      onPress: () => console.log('Cancel Pressed'),
+                      style: 'cancel',
+                    },
+                    {
+                      text: 'Delete',
+                      onPress: () =>
+                        dispatch(tripActions.deleteTrip(itemData.item.id)),
+                    },
+                  ],
+                  {cancelable: true},
+                );
               }}>
               <Text
                 style={{fontWeight: 'bold', fontSize: 16, color: '#FFFFFF'}}>
